+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Unreleased
 ### Changed
 - Deprecate default instantiation of TestUtils in TestCase - `$this->utils` will be removed in next major release. Do your own instantiation when needed.
+- Deprecate `TestUtils::setSelect2Value()` method. Use directly the new `Select2` component.
 
 ## 2.0.1 - 2016-12-01
 ### Fixed